+++ assets/scripts/ie-fix.js
@@ -1,4 +1,4 @@
-/* FIX FROM
+/* FROM
 /------------------------*/
 if (!Array.from) {
   Array.from = (function () {
@@ -76,3 +76,8 @@ if (!Array.from) {
     };
   }());
 }
+
+
+/* CSS PROPERTIES
+/------------------------*/
+window.MSInputMethodContext && document.documentMode && document.write('<script src="https://cdn.jsdelivr.net/gh/nuxodin/ie11CustomProperties@4.1.0/ie11CustomProperties.min.js"><\x2fscript>');