+++ classes/prefix_WPgutenberg/blocks/posts/inspector.js
@@ -153,7 +153,7 @@ export default class Inspector extends Component {
         const terms = posts[0]._links["wp:term"];
         Object.entries(terms).forEach(([key, value]) => {
           // get taxonomies
-          const getTaxvalues = select( 'core' ).getEntityRecords( 'taxonomy', value.taxonomy );
+          const getTaxvalues = select( 'core' ).getEntityRecords( 'taxonomy', value.taxonomy, { per_page: -1 } );
           let sectionOptions = [];
           let selectedOptions = [];
           if(getTaxvalues){