+++ classes/prefix_WPgutenberg/blocks/vimeo/attributes.js
@@ -26,6 +26,34 @@ const attributes = {
   videoDimensionY: {
     type: 'string',
     default: '3',
+  },
+  videoTableOfContent: {
+    type: 'boolean',
+    default: false,
+  },
+  videoTOCtoggle: {
+    type: 'boolean',
+    default: false,
+  },
+  videoTOCposition: {
+    type: 'string',
+    default: 'left'
+  },
+  videoTOCautoplay: {
+    type: 'boolean',
+    default: false,
+  },
+  videoTOCstop: {
+    type: 'boolean',
+    default: false,
+  },
+  videoLink: {
+    type: 'string',
+    default: '',
+  },
+  videoLinkTarget: {
+    type: 'string',
+    default: '_self',
   }
 }