+++ search.php
@@ -31,7 +31,7 @@ get_header();
   <?php endif; ?>
 </section>
 <nav class="pagination">
-  <?
+  <?php
   $big = 999999999; // need an unlikely integer
   echo paginate_links( array(
       'base' => str_replace( $big, '%#%', esc_url( get_pagenum_link( $big ) ) ),